requirements-updated.txt
# Core Flask and Extensions
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SocketIO==5.5.1
Flask-Bootstrap==3.3.7.1
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.3.1
gunicorn==20.1.0
itsdangerous==2.1.2
Werkzeug==2.2.3
Jinja2==3.1.2
eventlet==0.39.1
gevent==22.10.2

# API Services
openai==0.28.1
elevenlabs==0.2.24
deepgram-sdk==3.0.0
anthropic==0.49.0
requests==2.31.0
python-dotenv==1.0.0
dnspython==2.4.2

# Database and ORM
SQLAlchemy==2.0.7
alembic==1.10.3

# Auth and Security
Authlib==1.2.0
bcrypt==4.0.1
pyOpenSSL==23.1.1
cryptography==40.0.1

# CLI and System Utilities
click==8.1.8
psutil==5.9.5
colorlog==6.7.0

# Testing
pytest==7.3.1
pytest-flask==1.2.0

# Misc
pytz==2023.3
Pillow==10.0.0

# New additions from the code block
colorama==0.4.6
tqdm==4.67.1
Deprecated==1.2.18
tenacity==9.1.2

# New API and HTTP services
urllib3==2.3.0
aiohttp==3.11.16
httpx==0.28.1

# New AI and ML services
tiktoken==0.9.0
numpy==2.1.3
faiss-cpu==1.10.0
scikit-learn==1.6.1

# New audio processing services
pydub==0.25.1
ffmpeg-python==0.2.0
SoundFile==0.13.1
librosa==0.10.1
webrtcvad==2.0.10

# New WebSockets and real-time services
python-socketio==5.13.0
python-engineio==4.12.0
simple-websocket==1.1.0
websockets==15.0.1 